# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ASPICE Documentation Platform API",
    "description": "Multi-tenant, role-based SaaS API for automating automotive ASPICE / ISO 26262 documentation from C++ source code repositories.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/projects/{project_id}/jobs": {
      "post": {
        "tags": ["jobs"],
        "summary": "Start an analysis job (admin only)",
        "operationId": "startJob",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StartJobRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job_id": { "type": "string" },
                    "status": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs/current": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get the current active job for a project",
        "operationId": "getCurrentJob",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" }
        ],
        "responses": {
          "200": {
            "description": "Current job or null",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job": {
                      "oneOf": [
                        { "$ref": "#/components/schemas/JobView" },
                        { "type": "null" }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs/{job_id}": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get job by ID",
        "operationId": "getJob",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/JobId" }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job": { "$ref": "#/components/schemas/JobView" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs/{job_id}/cancel": {
      "post": {
        "tags": ["jobs"],
        "summary": "Cancel a job (admin only)",
        "operationId": "cancelJob",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/JobId" }
        ],
        "responses": {
          "200": {
            "description": "Cancelled job",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job": { "$ref": "#/components/schemas/JobView" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs/{job_id}/resume": {
      "post": {
        "tags": ["jobs"],
        "summary": "Resume a paused job (admin only)",
        "operationId": "resumeJob",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/JobId" }
        ],
        "responses": {
          "200": {
            "description": "Resumed job",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job": { "$ref": "#/components/schemas/JobView" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs/{job_id}/events": {
      "get": {
        "tags": ["jobs"],
        "summary": "Server-Sent Events stream for live job progress",
        "operationId": "jobEvents",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/JobId" }
        ],
        "responses": {
          "200": {
            "description": "SSE stream. Events: phase_update, activity_update, log_line, job_complete, job_failed.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs/{job_id}/functions": {
      "get": {
        "tags": ["jobs"],
        "summary": "List C++ functions discovered by a job",
        "operationId": "listJobFunctions",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/JobId" }
        ],
        "responses": {
          "200": {
            "description": "Function list with visibility summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "functions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/FunctionView" }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "hidden": { "type": "integer" },
                        "new_since_last": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs/{job_id}/reexport": {
      "post": {
        "tags": ["jobs"],
        "summary": "Re-queue DOCX export for an existing job (admin only)",
        "operationId": "reexportJob",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/JobId" }
        ],
        "responses": {
          "200": {
            "description": "Re-export queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "job_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Project identifier"
      },
...
      "JobId": {
        "name": "job_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Analysis job identifier"
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict — resource already exists or state prevents the action",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "NOT_FOUND" },
              "message": { "type": "string" },
              "status": { "type": "integer", "example": 404 }
            }
          }
        }
      },
...
      "StartJobRequest": {
        "type": "object",
        "required": ["commit_sha"],
        "properties": {
          "commit_sha": { "type": "string" },
          "version_tag": { "type": "string", "nullable": true },
          "reference_version_id": { "type": "string", "nullable": true },
          "pause_after_phase1": { "type": "boolean", "default": false },
          "layer_filter": { "type": "string", "nullable": true }
        }
      },
      "AnalysisPhaseView": {
        "type": "object",
        "properties": {
          "number": { "type": "integer" },
          "name": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "running", "done", "failed"] },
          "duration_seconds": { "type": "integer", "nullable": true }
        }
      },
      "JobView": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "status": { "type": "string", "enum": ["queued", "running", "paused", "complete", "failed", "cancelled"] },
          "phase": { "type": "integer" },
          "phase_pct": { "type": "integer" },
          "current_activity": { "type": "string" },
          "activity_detail": { "type": "string" },
          "elapsed_seconds": { "type": "integer" },
          "eta_seconds": { "type": "integer", "nullable": true },
          "phases": { "type": "array", "items": { "$ref": "#/components/schemas/AnalysisPhaseView" } },
          "commit_sha": { "type": "string" },
          "branch": { "type": "string" },
          "version_id": { "type": "string", "nullable": true },
          "started_at": { "type": "string", "format": "date-time" },
          "completed_at": { "type": "string", "format": "date-time", "nullable": true },
          "error_message": { "type": "string", "nullable": true }
        }
      },
      "FunctionView": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "file_path": { "type": "string" },
          "layer": { "type": "string" },
          "group": { "type": "string" },
          "is_visible": { "type": "boolean" },
          "is_new": { "type": "boolean" },
          "description": { "type": "string" }
        }
      },
...
    }
  }
}